// Client/projects/coi/src/app/configuration/form-builder-create/form-builder-create-interface.ts
export interface component {
  componentId: number
  sectionId: number
  formBuilderId: any
  componentType: string
  componentOrder: number
  componentData: string
  componentRefId: string
  componentDescription: string
  componentHeader: string
  componentFooter: string
  isActive: string
  updateTimestamp: string
  updateUser: string
  isMandatory: any
  validationMessage: any
  label: any
  componentTypeDescription:any
  }

  export interface formHeader {
    formBuilderId: number
    formBuilderNumber: string
    versionNumber: number
    versionStatus: string
    title: string
    description: string
    isActive: string
    createTimestamp: string
    createUser: string
    updateTimestamp: string
    updateUser: string
    usages: any
    sections: any
  }

  export interface NewSection {
    sectionId: number
    formBuilderId: number
    sectionName: string
    sectionOrder: number
    sectionBusinessRule: any
    sectionDescription: string
    sectionHelpText: string
    sectionHeader: string
    sectionFooter: string
    isActive: string
    updateTimestamp: string
    updateUser: string
    sectionComponent: any
  }
// ...
  export interface ElementTree {
    componentTypeCode: string
    description: string
    isActive: string
    updateTimestamp: string
    updateUser: string
  }
// ...
export class  CreateComponentObject {
  sectionId: string;
  formBuilderId: string;
  componentType: any;
  componentOrder: number;
  componentData: string;
  componentRefId: string;
  description: string;
  componentHeader: string;
  componentFooter: string;
  isActive: string;
  componentTypeDescription: any;
  label:any;
  isMandatory:any;
  validationMessage:any
}

export class FormSectionObject {
    sectionId: any;
    formBuilderId: any;
    sectionName: any;
    sectionOrder: any;
    sectionBusinessRule: any;
    sectionDescription: any;
    sectionHelpText: any;
    sectionHeader: any;
    sectionFooter: any;
    isActive: any;
}